Fix for CustomButton click events. Fix for Widget layoutData/DockLayoutPanel issue.
The CustomButton wouldn't trigger clickHandlers that had been
registered. This would cause it and subclasses to not use the
clickHandlers. Fix the issue by triggering the proper event type. Left
in place the CustomEvent for back-compatability.

Widget has an issue when adding to a DockLayoutPanel. Related to
different scope rules b/t Java and Dart. Refactored layoutData var name
to fix the issue.